Nearly ok!
This is a nearly ok configuration of almost the latest updates of most of the artifacts for bindings.

The big thing needed here is to step back on the firebase-ml things a bit so we find a version which does not incur androidx bindings as we aren't quite ready for those. (17.x doesn't need it).

Also, there are some new/updated bindings we need
 - io.grpc
  - grpc-android
  - grpc-okhttp
  - grpc-protobuf-lite
  - grpc-stub
 - tensorflow lite to 1.12
 - okhttp3 (might be on nuget already)
 - okhttp update to 2.7.5 (might be on nuget already)
 - com.google.auto.value-auto-value-annotations